SecurityPkg: Add Tpm2DebugLib to Support Detailed Logging
The TPM code is currently very noisy (e.g. in a sample
platform, 4,000 of the 5,700 lines printed to the serial
port at DEBUG_INFO level were from the TPM code). For TPM
debugging, this is very critical information, but for
most builds it simply spams the logs and slows down the
build.

This commit adds Tpm2DebugLib which is used to control
TPM driver/library logging at three levels: Null, Simple,
and Verbose. It is expected that platforms will use
the Null or Simple library instances except when
debugging the TPM operation. This commit updates
Tpm2DeviceLibDTpm to use Tpm2DebugLib.

This design pattern was chosen to allow platform flexibility
because of the extremely large output of the TPM code. Simply
moving this to DEBUG_VERBOSE or defining a new error level bit
does not give a platform enough granularity; in addition,
there is a lot of code to just support dumping (> 700 lines)
that needs to live somewhere and can be saved from a binary
(including all the strings).

This commit also adds Tpm2DebugLib as a dependency in the
couple of DSCs that use one of the libraries/driver that
needs it.

After this patchset, using Tpm2DebugLibNull reduces a given
platform's serial output from 5,700 lines to 1,700.

Continuous-integration-options: PatchCheck.ignore-multi-package

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>
Co-authored-by: Bret Barkelew <brbarkel@microsoft.com>