ArmPkg: StMM: Use x29 register to store stack address
Now that the use of 18 registers for FF-A is supported. Registers
from x0-x17 can be used by FF-A to return call results. StMM also uses
registers x19-x23 to store SPMC boot information.

Use the x29 register to store the stack address instead of x12
to avoid it being clobbered by FF-A calls.

The x29 is normally used as a Frame Pointer Register, but it's not
used as such in this context, so it could be used here.

Signed-off-by: Mohamed Gamal Morsy <mohamed.morsy@arm.com>