MdeModulePkg: Set Page Writeable When Freeing Pages
Currently, if the DebugClearMemory bit is set in
the PcdDebugPropertyMask, CoreConvertPagesEx will
attempt to write a pattern to the pages being freed.
However, it does not check that the page is writeable,
which will cause a page fault if not. Furthermore,
if NX protections are not enabled, the core does
not ensure that any freed pages are RW, which is the
state expected when they are allocated next. If they
are not RW, the allocating driver will crash trying
to use them.

This patch updates the page freeing code to remove
EFI_MEMORY_RO and EFI_MEMORY_RP from pages that are
being freed. It does this through the GCD to keep it
in sync with the page table state. Pool Pages do not
have this applied because the GCD allocates and
frees pool memory on every operation, which could
cause a recursive call into the GCD, tripping the lock.
Page attributes should never be set on pool memory
(except by the core), because it affects the entire
page, which may have different owners of other pools, so
we expect pool pages to always be writeable.

Because a memory range may span GCD descriptors, FreePages
will attempt to free memory on a GCD descriptor level basis
(to preserve other attributes, such as cacheability
attributes) and if any range fails, let that memory leak
(returning an error to the caller) and attempt to free
the rest of the memory. This was deemed the best approach to
let memory that can't be guaranteed to be RW leak instead
of letting a driver crash when allocating it.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>